+++ valhalla/mjolnir/graphtilebuilder.h
@@ -504,8 +504,9 @@ public:
    *
    * @param edge_id  the edge id to modify
    * @param landmark the landmark to associate to the edge
+   * @return Returns true if the input landmark is indeed inserted in to the edge.
    */
-  void AddLandmark(const baldr::GraphId& edge_id, const Landmark& landmark);
+  bool AddLandmark(const baldr::GraphId& edge_id, const Landmark& landmark);
 
   /**
    * Is there an opposing edge with matching edgeinfo offset. The end node of the directed edge